+++ common/src/main/kotlin/com/cobbledex/jei/PokemonIngredientRenderer.kt
@@ -1,11 +1,9 @@
 package com.cobbledex.jei
 
-import com.cobbledex.DiscoveryAliases
 import com.cobbledex.PokemonItemCache
 import com.cobbledex.SpeciesNameNormalizer
 import com.cobbledex.SpawnDisplayHelper
 import mezz.jei.api.ingredients.IIngredientRenderer
-import net.minecraft.ChatFormatting
 import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.network.chat.Component
 import net.minecraft.world.item.TooltipFlag
@@ -26,12 +24,7 @@ class PokemonIngredientRenderer : IIngredientRenderer<PokemonIngredient> {
     }
 
     override fun getTooltip(ingredient: PokemonIngredient, tooltipFlag: TooltipFlag): List<Component> {
-        val lines = SpawnDisplayHelper.buildPokemonTooltipLines(ingredient.species, ingredient.displayName).toMutableList()
-        val searchText = DiscoveryAliases.pokemonSearchText(ingredient.species)
-        if (searchText.isNotBlank()) {
-            lines.add(Component.literal(searchText).withStyle(ChatFormatting.BLACK))
-        }
-        return lines
+        return SpawnDisplayHelper.buildPokemonTooltipLines(ingredient.species, ingredient.displayName)
     }
 
     fun canRender(species: String): Boolean = PokemonItemCache.canRender(species)